Pulizie_BOT_MOVE/requirements.txt
# Bot Telegram Gestione Pulizie
# Requirements / Dipendenze

# Telegram Bot
python-telegram-bot==20.7

# HTTP Requests (per Google Maps API)
requests==2.31.0

# Geolocalizzazione e calcoli GPS
geopy==2.4.1

# Export Excel
openpyxl==3.1.2

# File Locking (per protezione Excel da race conditions)
filelock==3.13.1

# Utility
python-dateutil==2.8.2
